+++ app/components/dashboard/analysis/Counters.vue
@@ -56,7 +56,7 @@ onMounted(async () => {
         />
       </CardHeader>
       <CardContent>
-        <NumberFlow class="text-2xl font-bold tabular-nums" :class="{ 'opacity-60 blur-md': !counters.visits }" :value="counters.visits" />
+        <NumberFlow class="text-2xl font-bold tabular-nums" :value="counters.visits" />
       </CardContent>
     </Card>
     <Card class="gap-0">
@@ -69,7 +69,7 @@ onMounted(async () => {
         <Users aria-hidden="true" class="h-4 w-4 text-muted-foreground" />
       </CardHeader>
       <CardContent>
-        <NumberFlow class="text-2xl font-bold tabular-nums" :class="{ 'opacity-60 blur-md': !counters.visitors }" :value="counters.visitors" />
+        <NumberFlow class="text-2xl font-bold tabular-nums" :value="counters.visitors" />
       </CardContent>
     </Card>
     <Card class="gap-0">
@@ -82,7 +82,7 @@ onMounted(async () => {
         <Flame aria-hidden="true" class="h-4 w-4 text-muted-foreground" />
       </CardHeader>
       <CardContent>
-        <NumberFlow class="text-2xl font-bold tabular-nums" :class="{ 'opacity-60 blur-md': !counters.referers }" :value="counters.referers" />
+        <NumberFlow class="text-2xl font-bold tabular-nums" :value="counters.referers" />
       </CardContent>
     </Card>
   </div>